Move query selectors back to domUpdates
